pyproject.toml
[tool.poetry]
name = "particle-pattern"
version = "0.1.0"
description = "Creating patterns with particle simulations."
authors = ["gordonbchen <gordonbchen@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.1"
pygame = "^2.6.1"
scipy = "^1.14.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
